[OVN] Set send_periodic to False on provider networks
This patch adds a check prior to setting the send_periodic configuration
option for the LRP's ipv6_ra_configs to see if the network the LRP is
connected to is a provider network and the port is a gateway port.

In case it is a provider network and the port is a gateway port,
send_periodic should be set to False to avoid leaking the RAs generated
by ovn-controller to the tenant networks onto the provider network.

Closes-Bug: #1919347
Change-Id: I4e1ffb3a1eea147a750def245f869aca9e036b88
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>