Set proxy_delay to one when using proxy ARP
Neutron DVR uses proxy ARP in fip namespaces to respond
to ARP requests for instance floating IPs. In doing so
it is susceptible to a random delay of up to (by
default) 800ms which is added to the time taken to
respond to ARP requests. This causes an initial delay
to ARP reponses that is entirely avoidable by changing this
parameter to one, instead of the default, to make it as
short as possible.

NOTE: Setting this to zero is actually undefined and will
cause the kernel to choose a random delay from 0 to
U32_MAX so is not advised. Gleaned from this comment in
__get_random_u32_below(), which is eventually called
from pneigh_enqueue():

/*
 * This function is technically undefined for ceil == 0, and in fact
 * for the non-underscored constant version in the header, we build bug
 * on that. But for the non-constant case, it's convenient to have that
 * evaluate to being a straight call to get_random_u32(), so that
 * get_random_u32_inclusive() can work over its whole range without
 * undefined behavior.
 */

Will propose a kernel change to fix this but cannot
assume it will be in a distro kernel for a while.

Change-Id: I0dc65b17ef436a97d0fcbd164d124ec59a1b2797
Closes-Bug: #1920975