Check there are no min BW QoS in "BaseOVSTestCase"
This patche adds a check in "BaseOVSTestCase" setUp method. Each
time a test is executed, the test case checks there are no
minimum bandwidth QoS/queues related to this OVS instance. The OVS
instance generates a random "_min_bw_qos_id" in the initialization.
This ID is assigned to the OVS QoS registers, in "external_ids:id".

Related-Bug: #1922919

Change-Id: I6955aadc51bfc9181ec84f7d54a219ffa3fd9115