Always use absolute path for custom kill-scripts
When rootwrap was used to run custom kill-scripts, it worked fine with
the executable name without full path as path was in the rootwrap's
exec_dirs config option thus rootwrap was able to find it.

But as we migrated to privsep it don't works like that anymore.
It's better to use absolute path of the kill-script always and that
patch changes to do it like that.

Change-Id: I66d70304530935e2add3345aba6aa3c549a0a2df
Closes-Bug: #1923198