[OVN] External network ports (SR-IOV) QoS is handled by SR-IOV agent
This commit is similar to [1]. The OVN QoS extension does not manage
the external type ports (SR-IOV); the port QoS policy of those ports
is handled by the SR-IOV agent QoS extension.

When a network is updated, the OVN QoS extension should check the type
of port bound. In case of having an "external" port (SR-IOV), the QoS
extension should not handle it.

[1]https://review.opendev.org/c/openstack/neutron/+/780054

Closes-Bug: #1925218

Change-Id: Idf85e304897c9d1073091c6cabb86be39640f332