Don't ever give up trying to connect to OVN DBs
It doesn't really make since to give up connectimg to the OVN DBs
since we can't do anything without them. If ovsdb-server is slow
and takes more than timeout seconds, everything reconnecting after
partial downloads and starting over is not going to make things
better.

We can change the behavior in ovsdbapp, but doing it without
making non-backportable API changes isn't easily doable.
The plan would be, after merging this back through stable, to
modify ovsdbapp to allow setting different connection and txn
timeouts and being able to disable the timeout in wait_for_change
with a value of -1. Then in the main branch of neutron we can
use that going forward.

Change-Id: Ia9e23113fdeebf0b99085da200c3d61b71567d36