Add missing "system" scope in some of the new API policies
During the migration to the new RBAC policies we made mistakes and we
allowed some of the APIs to be available for SYSTEM users but we allowed
it only for "project" scope.
As we discussed during the PTG, policy in such case should allows to use
such API also for the "system" scope tokens.
This patch adds it where it was still missing.

Partially-Implements blueprint: secure-rbac-roles

Change-Id: Icfbd00040834b311e78138cc93e7ab86355e557c