Do not include dynamically loaded values into OVO testing
Those OVO values loaded dynamically (for example [1]) should not be
added to the tested object. A dynamic column in a DB register is a
``sqlalchemy.orm.AppenderQuery`` object that represents the needed
query to retrieve this value.

These OVO fields, mapped from the DB object, should be considered as
synthetic fields and not be included in
``_BaseObjectTestCase.obj_fields``.

[1]https://github.com/openstack/neutron/blob/ff2464bf33f1c84bd8525eca68c5774f5e14f1f1/neutron/db/models/securitygroup.py#L103-L106

Closes-Bug: #1926399

Change-Id: I4b59dbc71316eede52b22c0788de5fd682fbff30