[OVN] Add baremetal support with Neutron DHCP agent
This patch now creates OVN "external" ports for Neutron ports with
VNIC_BAREMETAL. This ports will be scheduled on the OpenStack Controller
nodes (or OVN Gateway nodes) and are responsible for replying to the
ARP requests coming from the baremetal nodes.

This patch also disables OVN's built-in DHCP server for VNIC_BAREMETAL
ports. This is because OVN DHCP server does not yet fully support
chainloading from PXE to iPXE, this feature is work-in-progress right
now. A following patch to this one will be sent in the future adding
support for OVN's built-in DHCP server being used with baremetal nodes.

This patch implements the "Part 1" from bug #1971431.

Partial-bug: #1971431
Change-Id: I6b234fbe1b7c54b41a1b8b430fdf0ac76993af96
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>