Update Neutron's Liuetenants
This patch updates list of the Neutron's stadium's projects lieutenants
and the list of the bugs' contact persons.

In details this patch:
- sets Rodolfo Alonso Hernandez as contact person for db and qos
  related issues,
- adds Oleg Bondarev as "loadimpact" bugs,
- removes Matt Riedemann as contact person for "logging" bugs,
- sets PTL/Drivers team as conctact for "troubleshooting" related bugs,

It also sets Lajos Katona as "Testing" lieutenant.

Finally it removes networking-ovn and neutron-fwaas from the list of
stadium projects and removes tag "fwaas" from the list of bug tags.
Neutron-fwaas was deprecated and isn't part of the stadium since long
time and networking-ovn is now one of the Neutron in-tree drivers.

Change-Id: Id4b928e077ed684c67d4b5054f12653d63f70788