HA-non-DVR router don't need manually add static route
When a router set as HA mode, The keepalived process will take over
the route entry's generation. So, the codes that add static route
is redundant.

But, for DVR-HA router, in dvr_snat node the keepalived process run
in snat-namespace and don't take over qrouter-namespace, so the
manually add static route codes still need be called.

Closes-Bug: #1927849
Change-Id: Id09de6c43c0fab4009336e253c88f54219398053