Install "pyroute2" as a doc job depedency
By adding this library to the doc requirements file, tox will install
it limiting the upper version according to
"requirements/upper-constraints.txt".

"pyroute2" is called in tox doc job as mentioned in [1]. Since version
0.6.0, the library structure changed. To keep backwards compatibility,
all modules are imported from the main namespace. As reported in the
referred bug, the library "mitogen" must be installed first.

In stable releases that use older "pyroute2" versions, this library is
not needed but this patch should enforce the upper library version.

Related-Bug: #1928913

[1]https://github.com/openstack/neutron/blob/2a4074918c248adf84b0afa404d30c18710da4b2/doc/source/conf.py#L33-L40

Change-Id: I0cf458f4532d313e593a943653e8310b93ac1e9c