Move dns-integration extension to the ML2_SUPPORTED_API_EXTENSIONS list
dns-integration extension is actually implemented as ML2 extension
driver, so has nothing to do with the L3 plugin.
In OVN it was listed in the ML2_SUPPORTED_API_EXTENSIONS_OVN_L3 but it
should be in ML2_SUPPORTED_API_EXTENSIONS. This patch moves it to the
correct list.
This patch also removes disable_dns_extension_by_extension_drivers
method from the OVN L3 plugin class as there is no need to disable this
extension by L3 agent.

Related-Bug: #1929676
Change-Id: I456afcc9054627e74b18460d169052f262451c22