Remove tox_install_siblings=False from the functional job's definition
It is set to true by default in base jobs. Also e.g. in
functional job for neutron-dynamic-routing it is set to True so
it should be fine to use default value in the neutron-functional job as
well.

Related-bug: #1932483
Change-Id: I2c8392628dd63ca612c6d8428daab66724e88c20
(cherry picked from commit dbfa1b5d9298dc146ff5eb6e560fea42600cf64f)