Add devstack local.conf sample for ML2 OVS
Since devstack had set OVN as the default backend for Neutron.
Then the minimum local.conf [1] for ML2 ovs will not work at
all. For some local testing of ML2 OVS, it is not right deployment
for users to test the ML2 OVS related cases.

This patch adds a sample local.conf for ml2 ovs to install a small
all in one environment for Neutron testing.

Sample tested OS:
1. CentOS Stream 8
2. CentOS Linux 8

[1] https://docs.openstack.org/devstack/latest/#create-a-local-conf

Closes-Bug: #1934466
Change-Id: Ie7bac1d2819c332a94a0ff308a300638c17f1b1f