[DVR] Fix update of the MTU in the DVR HA routers
This is follow up of the patch [1] which fixed updating MTU in the
snat namespace for the DVR routers.
In case of DVR-HA routers there was additional issue with that as
L3 agent tried to update MTU for the qr- interface in the
self.ha_namespace which, for DVR-HA routers is snat namespace.

This patch fixes that issue by setting MTU on the qr- interface in
qrouter namespace and also setting MTU on the snat interface in snat
namespace.

[1] https://review.opendev.org/c/openstack/neutron/+/799226

Conflicts:
    neutron/tests/functional/agent/l3/test_dvr_router.py

Closes-bug: #1933273
Change-Id: I409bc674b65e4f495ebd42d03e97a09d51482339
(cherry picked from commit 21eabbcf03cb29be9899adbd6eaea207892cf07b)