Remove orphaned event_lock_name attribute
After Hash Ring support was added, this event_lock_name attribute
was no longer used by anything. The ovn-octavia-provider sets it
in its subclass and then calls set_lock on it, which is what used
to happen in this tree. It should be safe to remove. Currently it
is confusing because it looks like subclasses of OvnIdl involve
ovsdb locking.

Change-Id: I5231d464cddfb9fa201f3762d2c9391ed037f0a5